# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/image": {
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "upload image",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "example": "image.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image": {
                                            "type": "string",
                                            "example": "image/AHlaX9L2WeivBN17QNXdtXfagLkLhKPj8k4MhaA6.png"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "14"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "14"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-image/{image}": {
            "delete": {
                "tags": [
                    "Image"
                ],
                "summary": "delete image",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "ID Image",
                        "example": 50
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "you deleted your image"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-image": {
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "update image",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "example": "image.png"
                                    },
                                    "id": {
                                        "type": "integer",
                                        "example": "52"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You changed your image"
                                        },
                                        "image": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "52"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "image/AHlaX9L2WeivBN17QNXdtXfagLkLhKPj8k4MhaA6.png"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}